@@ -13,7 +13,7 @@
     "ldm": "./bin/ldm"
   },
   "scripts": {
-    "prepublish": "npm run compile",
+    "prepublish": "npm run clean && npm run compile",
     "compile": "tsc",
     "watch": "tsc -w",
     "clean": "rimraf lib"
@@ -22,17 +22,19 @@
     "@types/chokidar": "^1.6.0",
     "@types/gulp": "^4.0.3",
     "@types/gulp-newer": "0.0.30",
+    "@types/gulp-sourcemaps": "0.0.30",
     "@types/node": "^7.0.22",
     "@types/through2": "^2.0.33",
     "@types/vinyl-paths": "0.0.29",
     "rimraf": "^2.6.1",
-    "typescript": "^2.3.3"
+    "typescript": "^2.4.1"
   },
   "dependencies": {
     "check-if-windows": "^1.0.0",
     "fs-extra": "^3.0.1",
     "gulp": "github:gulpjs/gulp#4.0",
     "gulp-newer": "^1.3.0",
+    "gulp-sourcemaps": "^1.6.0",
     "through2": "^2.0.3",
     "vinyl-paths": "^2.1.0"
   }
