@@ -4,61 +4,48 @@
   "version": "0.0.1-alpha.1",
   "license": "Apache-2.0",
   "dependencies": {
-    "monaco-css": "^1.3.1",
-    "monaco-editor": "^0.8.3",
-    "monaco-editor-core": "^0.8.2",
-    "monaco-html": "^1.2.1",
-    "monaco-json": "^1.2.2",
-    "monaco-languages": "^0.7.0",
-    "monaco-typescript": "^2.1.2",
-    "reflect-metadata": "^0.1.10",
-    "theia-core": "latest",
-    "yargs": "^8.0.1"
+    "theia-core": "latest"
   },
   "localDependencies": {
     "theia-core": "../.."
   },
   "scripts": {
     "bootstrap": "npm install ../../config/local-dependency-manager && npm run localinstall",
-    "localinstall": "ldm install --dev",
+    "build": "npm run build:localdeps && npm run build:frontend",
+    "build:frontend": "webpack --target web && cp src-gen/frontend/index.html lib",
+    "build:localdeps": "ldm run build && ldm sync --original-sources",
     "clean": "rimraf lib",
-    "prepare": "npm run clean && npm run build",
-    "build": "npm run build:localdeps && npm run build:backend && npm run build:frontend",
-    "build:backend": "tsc && npm run lint",
-    "build:frontend": "webpack && cp src/frontend/index.html lib/frontend",
-    "build:localdeps": "ldm run build && ldm sync",
-    "lint": "tslint -c ../../config/tslint/tslint.json --project ./tsconfig.json",
-    "start": "concurrently --names backend,webpack-server --prefix \"[{name}]\" \"npm run start:backend\" \"npm run start:frontend\"",
-    "start:backend": "node ./lib/backend/main.js | bunyan",
-    "start:backend:debug": "node ./lib/backend/main.js --loglevel=debug | bunyan",
-    "start:frontend": "webpack-dev-server --open",
     "cold:start": "npm run clean && npm run build && npm start",
-    "watch:prod": "concurrently --names watch-deps,watch-backend,watch-frontend --prefix \"[{name}]\" \"npm run watch:localdeps:prod\" \"npm run watch:backend\"  \"npm run watch:frontend\"",
-    "watch": "concurrently --names watch-deps,watch-backend,watch-frontend --prefix \"[{name}]\" \"npm run watch:localdeps\" \"npm run watch:backend\"  \"npm run watch:frontend\"",
-    "watch:backend": "tsc --watch",
-    "watch:frontend": "npm run build:frontend && webpack --watch",
-    "watch:localdeps:prod": "ldm watch --sync --run=watch:prod",
-    "watch:localdeps": "ldm watch --sync --run=watch",
-    "test:ui": "wdio wdio.conf.js"
+    "localinstall": "ldm install --dev --original-sources",
+    "prepare": "npm run clean && npm run build",
+    "start": "concurrently --names backend,webpack-server --prefix \"[{name}]\" \"npm run start:backend\" \"npm run start:frontend\"",
+    "start:backend": "node ./src-gen/backend/main.js | bunyan",
+    "start:backend:debug": "node ./src-gen/backend/main.js --loglevel=debug | bunyan",
+    "start:frontend": "webpack-dev-server --open",
+    "test:ui": "concurrently --names backendfrontend,uitest --prefix \"[{name}]\" \"npm run start\" \"wdio --waitforTimeout 300000 wdio.conf.js\"",
+    "watch": "concurrently --names watch-deps,watch-frontend --prefix \"[{name}]\" \"npm run watch:localdeps\" \"npm run watch:frontend\"",
+    "watch:frontend": "npm run build:frontend && webpack --target web --watch",
+    "watch:localdeps": "ldm watch --sync --run=watch --original-sources"
   },
   "devDependencies": {
     "@types/chai": "^3.5.1",
     "@types/webdriverio": "^4.7.0",
-    "@types/yargs": "^6.6.0",
     "bunyan": "^1.8.10",
     "chai": "^3.5.0",
+    "circular-dependency-plugin": "^2.0.0",
     "concurrently": "^3.4.0",
+    "copy-webpack-plugin": "^4.0.1",
     "css-loader": "^0.28.1",
     "file-loader": "^0.11.1",
+    "font-awesome-webpack": "0.0.5-beta.2",
+    "less": "^2.7.2",
     "local-dependency-manager": "../../config/local-dependency-manager",
     "rimraf": "^2.6.1",
     "selenium-standalone": "^6.2.0",
     "source-map-loader": "^0.2.1",
-    "ts-loader": "^2.0.3",
     "ts-node": "^3.0.2",
     "tslint": "^4.5.1",
-    "tslint-loader": "^3.4.3",
-    "typescript": "^2.3.2",
+    "typescript": "^2.4.1",
     "url-loader": "^0.5.8",
     "wdio-mocha-framework": "^0.5.9",
     "wdio-phantomjs-service": "^0.2.2",
