@@ -55,33 +65,19 @@
   "devDependencies": {
     "@types/chai": "^3.4.35",
     "@types/chai-as-promised": "0.0.29",
-    "@types/chokidar": "^1.6.0",
-    "@types/karma": "^0.13.33",
     "@types/mocha": "^2.2.39",
     "chai": "^3.5.0",
     "chai-as-promised": "^6.0.0",
-    "circular-dependency-plugin": "^2.0.0",
-    "copy-webpack-plugin": "^4.0.1",
-    "css-loader": "^0.26.2",
-    "file-loader": "^0.11.1",
     "find-package-json": "^1.0.0",
     "gulp": "^3.9.1",
     "gulp-decompress": "^2.0.1",
     "gulp-download": "0.0.1",
     "mocha": "^3.2.0",
     "rimraf": "^2.6.1",
-    "source-map-loader": "^0.1.6",
-    "style-loader": "^0.13.2",
-    "ts-loader": "^2.0.1",
-    "ts-node": "^2.1.0",
-    "tsc-watch": "^1.0.5",
+    "ts-node": "^3.1.0",
     "tslint": "^4.5.1",
-    "typescript": "^2.3.2",
-    "url-loader": "^0.5.8",
-    "uuid": "^3.0.1",
-    "webpack": "^2.2.1",
-    "webpack-dev-server": "^2.4.1",
-    "webpack-merge": "^3.0.0"
+    "typescript": "^2.4.1",
+    "uuid": "^3.0.1"
   },
   "nyc": {
     "include": [
