@@ -5,30 +5,37 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-import { Endpoint } from '../../application/common/endpoint';
+import { inject, injectable } from "inversify";
+import { Endpoint, Disposable } from '../../application/common';
+import { Widget, BaseWidget, Message } from '../../application/browser';
 import { WebSocketConnectionProvider } from '../../messaging/browser';
-import { Disposable } from '../../application/common';
-import { DisposableCollection } from 'vscode-ws-jsonrpc/lib/disposable';
-import { Message } from '@phosphor/messaging/lib';
-import { Widget } from '@phosphor/widgets/lib';
+import { WorkspaceService } from "../../workspace/browser";
 import * as Xterm from 'xterm';
 import 'xterm/lib/addons/fit/fit';
 import 'xterm/lib/addons/attach/attach';
 
 let num = 0
 
-export class TerminalWidget extends Widget implements Disposable {
+export const TerminalWidgetFactory = Symbol('TerminalWidgetFactory');
+export interface TerminalWidgetFactory {
+    (): TerminalWidget;
+}
+
+@injectable()
+export class TerminalWidget extends BaseWidget {
 
     private pid: string | undefined
     private term: Xterm
     private cols: number = 80
     private rows: number = 40
-    private disposables = new DisposableCollection()
     private endpoint: Endpoint
 
-    constructor(private websocketConnectionProvider: WebSocketConnectionProvider) {
-        super()
-        this.endpoint = new Endpoint({ path: '/terminals' })
+    constructor(
+        @inject(WorkspaceService) protected readonly workspaceService: WorkspaceService,
+        @inject(WebSocketConnectionProvider) protected readonly webSocketConnectionProvider: WebSocketConnectionProvider
+    ) {
+        super();
+        this.endpoint = new Endpoint({ path: '/services/terminals' })
         num++
         this.id = 'terminal-' + num
         this.title.caption = 'Terminal ' + num
@@ -49,14 +56,12 @@ export class TerminalWidget extends Widget implements Disposable {
         this.registerResize()
         this.startNewTerminal()
 
-        this.disposables.push({
-            dispose: () => {
-                this.term.destroy()
-            }
-        })
+        this.toDispose.push(Disposable.create(() =>
+            this.term.destroy()
+        ));
     }
 
-    protected registerResize() {
+    protected registerResize(): void {
         let initialGeometry = (this.term as any).proposeGeometry()
         this.cols = initialGeometry.cols;
         this.rows = initialGeometry.rows;
@@ -73,51 +78,40 @@ export class TerminalWidget extends Widget implements Disposable {
         (this.term as any).fit()
     }
 
-    protected startNewTerminal() {
-        fetch(this.endpoint.getRestUrl().toString() + '?cols=' + this.cols + '&rows=' + this.rows, { method: 'POST' }).then((res) => {
-            res.text().then((pid: string) => {
-                this.pid = pid;
-                let socket = this.createWebSocket(pid);
-                socket.onopen = () => {
-                    (this.term as any).attach(socket);
-                    (this.term as any)._initialized = true
-                }
-                socket.onclose = () => {
-                    this.title.label = `<terminated>`
-                };
-                socket.onerror = (err) => {
-                    console.error(err)
-                }
-                this.disposables.push({
-                    dispose() {
-                        socket.close()
-                    }
-                })
-            });
+    protected async startNewTerminal(): Promise<void> {
+        const root = await this.workspaceService.root;
+        const res = await fetch(this.endpoint.getRestUrl().toString() + '?cols=' + this.cols + '&rows=' + this.rows, {
+            method: 'POST',
+            headers: {
+                'Content-Type': 'application/json'
+            },
+            body: JSON.stringify({ uri: root.uri })
         });
+        this.pid = await res.text();
+        const socket = this.createWebSocket(this.pid);
+        socket.onopen = () => {
+            (this.term as any).attach(socket);
+            (this.term as any)._initialized = true;
+        };
+        socket.onclose = () => {
+            this.title.label = `<terminated>`;
+        };
+        socket.onerror = (err) => {
+            console.error(err);
+        };
+        this.toDispose.push(Disposable.create(() =>
+            socket.close()
+        ));
     }
 
     protected createWebSocket(pid: string): WebSocket {
-        const url = this.endpoint.getWebSocketUrl().appendPath(pid)
-        return this.websocketConnectionProvider.createWebSocket(url.toString(), { reconnecting: false })
-    }
-
-    dispose() {
-        if (this.isDisposed) {
-            return;
-        }
-        super.dispose()
-        this.disposables.dispose()
+        const url = this.endpoint.getWebSocketUrl().resolve(pid)
+        return this.webSocketConnectionProvider.createWebSocket(url.toString(), { reconnecting: false })
     }
 
     protected onActivateRequest(msg: Message): void {
         super.onActivateRequest(msg)
         this.term.focus()
-    }
-
-    protected onCloseRequest(msg: Message): void {
-        super.onCloseRequest(msg)
-        this.dispose()
     }
 
     private resizeTimer: any
