@@ -3,24 +3,21 @@ const paths = require('./paths');
 module.exports = function (dirname) {
     return [
         {
-            test: /\.tsx?$/,
-            exclude: [
-                paths(dirname).NODE_MODULES
-            ],
-            loader: 'ts-loader'
-        },
-        {
             test: /\.css$/,
             loader: 'style-loader!css-loader'
         },
         {
-            test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
+            test: /\.(ttf|eot|svg)(\?v=\d+\.\d+\.\d+)?$/,
             loader: 'url-loader?limit=10000&mimetype=image/svg+xml'
         },
         {
             test: /\.js$/,
             enforce: 'pre',
             loader: 'source-map-loader'
+        },
+        {
+            test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
+            loader: "url-loader?limit=10000&mimetype=application/font-woff"
         }
     ];
 };
