@@ -1,47 +1,8 @@
-const webpack = require("webpack");
-const path = require("path");
+/*
+ * Copyright (C) 2017 TypeFox and others.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ */
 
-module.exports = require("../../config/webpack/webpack.config")(__dirname, {
-    devServer: {
-        proxy: {
-            '/filesystem/*': {
-                target: 'ws://localhost:3000',
-                ws: true
-            },
-            '/languages/*': {
-                target: 'ws://localhost:3000',
-                ws: true
-            },
-            '/terminals/*': {
-                target: 'ws://localhost:3000',
-                ws: true
-            },
-            '/logger/*': {
-                target: 'ws://localhost:3000',
-                ws: true
-            },
-
-            '*': 'http://localhost:3000'
-        },
-        historyApiFallback: true,
-        hot: true,
-        inline: true,
-        stats: {
-            colors: true,
-            warnings: false
-        },
-        host: process.env.HOST || '127.0.0.1',
-        port: process.env.PORT
-    }, plugins: [
-        new webpack.HotModuleReplacementPlugin()
-    ],
-    resolve: {
-	// These shims are needed for bunyan
-	alias: {
-            'dtrace-provider': path.resolve(__dirname, './webpack_empty.js'),
-	    'safe-json-stringify': path.resolve(__dirname, './webpack_empty.js'),
-            mv: path.resolve(__dirname, './webpack_empty.js'),
-            'source-map-support': path.resolve(__dirname, './webpack_empty.js')
-	}
-    }
-});
+module.exports = require("theia-core/config/webpack/webpack.config.web.dev")(__dirname, 3000, "localhost");
