@@ -5,6 +5,8 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-export * from "./application";
-export * from "./application-module";
+export * from "./backend-application";
+export * from "./backend-application-module";
 export * from "./debug";
+export * from "./file-uri";
+export * from "./logger-backend-module";
