@@ -5,4 +5,4 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-export * from './backend-cpp-module';
+export * from './cpp-backend-module';
