@@ -4,5 +4,8 @@
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
-
-export * from './filesystem-client-module';
+export * from './filesystem-frontend-module';
+export * from './filesystem-commands';
+export * from './location';
+export * from './file-tree';
+export * from './file-dialog';
