@@ -5,7 +5,7 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-export * from './filesystem'
-export * from './filesystem-selection'
-export * from './filesystem-watcher'
+export * from './filesystem';
+export * from './filesystem-selection';
+export * from './filesystem-watcher';
 export * from './file-resource';
