@@ -5,19 +5,23 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-import { FileChangesEvent, FileChange, FileChangeType, FileSystemClient } from '../common';
-import * as chai from "chai";
-import * as chaiAsPromised from "chai-as-promised";
-import * as fs from "fs-extra";
-import * as os from "os";
+import * as chai from 'chai';
+import * as assert from 'assert';
+import * as chaiAsPromised from 'chai-as-promised';
+import * as fs from 'fs-extra';
+import * as os from 'os';
 import URI from "../../application/common/uri";
-import { FileUri } from "../../application/node/file-uri";
+import { Logger } from "../../application/common";
+import { FileUri } from "../../application/node";
 import { FileSystem } from "../common/filesystem";
+import { FileSystemWatcher, FileSystemWatcherClientListener, FileChange, FileChangeType } from '../common';
 import { FileSystemNode } from "./node-filesystem";
+import { ChokidarFileSystemWatcherServer } from './chokidar-filesystem-watcher';
+
 
 const expect = chai.expect;
 const uuidV1 = require('uuid/v1');
-const TEST_ROOT = FileUri.create(os.tmpdir()).appendPath("node-fs-root");
+const TEST_ROOT = FileUri.create(os.tmpdir()).resolve("node-fs-root");
 
 describe("NodeFileSystem", () => {
 
@@ -26,6 +30,7 @@ describe("NodeFileSystem", () => {
 
     let root: URI;
     let fileSystem: FileSystem;
+    let watcher: FileSystemWatcher;
 
     before(() => {
         chai.config.showDiff = true;
@@ -35,19 +40,19 @@ describe("NodeFileSystem", () => {
     });
 
     beforeEach(() => {
-        root = TEST_ROOT.appendPath(uuidV1());
+        root = TEST_ROOT.resolve(uuidV1());
         fs.mkdirsSync(FileUri.fsPath(root));
         expect(fs.existsSync(FileUri.fsPath(root))).to.be.true;
         expect(fs.readdirSync(FileUri.fsPath(root))).to.be.empty;
         roots.push(root);
         fileSystem = createFileSystem();
         fileSystems.push(fileSystem);
+
+        watcher = createFileSystemWatcher();
     });
 
     after(() => {
-        fileSystems.forEach(fileSystem => {
-            (fileSystem as any).watcher.close();
-        });
+        watcher.dispose();
         roots.map(root => FileUri.fsPath(root)).forEach(root => {
             if (fs.existsSync(root)) {
                 try {
@@ -62,14 +67,14 @@ describe("NodeFileSystem", () => {
     describe("01 #getFileStat", () => {
 
         it("Should be rejected if not file exists under the given URI.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             return fileSystem.getFileStat(uri.toString()).should.eventually.be.rejectedWith(Error);
         });
 
         it("Should return a proper result for a file.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             fs.writeFileSync(FileUri.fsPath(uri), "foo");
             expect(fs.statSync(FileUri.fsPath(uri)).isFile()).to.be.true;
 
@@ -80,8 +85,8 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should return a proper result for a directory.", () => {
-            const uri_1 = root.appendPath("foo.txt");
-            const uri_2 = root.appendPath("bar.txt");
+            const uri_1 = root.resolve("foo.txt");
+            const uri_2 = root.resolve("bar.txt");
             fs.writeFileSync(FileUri.fsPath(uri_1), "foo");
             fs.writeFileSync(FileUri.fsPath(uri_2), "bar");
             expect(fs.statSync(FileUri.fsPath(uri_1)).isFile()).to.be.true;
@@ -98,14 +103,14 @@ describe("NodeFileSystem", () => {
     describe("02 #resolveContent", () => {
 
         it("Should be rejected with an error when trying to resolve the content of a non-existing file.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             return fileSystem.resolveContent(uri.toString()).should.eventually.be.rejectedWith(Error);
         });
 
         it("Should be rejected with an error when trying to resolve the content of a directory.", () => {
-            const uri = root.appendPath("foo");
+            const uri = root.resolve("foo");
             fs.mkdirSync(FileUri.fsPath(uri));
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.true;
             expect(fs.statSync(FileUri.fsPath(uri)).isDirectory()).to.be.true;
@@ -114,7 +119,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should be rejected with an error if the desired encoding cannot be handled.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             fs.writeFileSync(FileUri.fsPath(uri), "foo", { encoding: "utf8" });
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.true;
             expect(fs.statSync(FileUri.fsPath(uri)).isFile()).to.be.true;
@@ -124,7 +129,7 @@ describe("NodeFileSystem", () => {
         })
 
         it("Should be return with the content for an existing file.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             fs.writeFileSync(FileUri.fsPath(uri), "foo", { encoding: "utf8" });
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.true;
             expect(fs.statSync(FileUri.fsPath(uri)).isFile()).to.be.true;
@@ -134,7 +139,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should be return with the stat object for an existing file.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             fs.writeFileSync(FileUri.fsPath(uri), "foo", { encoding: "utf8" });
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.true;
             expect(fs.statSync(FileUri.fsPath(uri)).isFile()).to.be.true;
@@ -159,7 +164,7 @@ describe("NodeFileSystem", () => {
     describe("03 #setContent", () => {
 
         it("Should be rejected with an error when trying to set the content of a non-existing file.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             const stat = {
@@ -171,7 +176,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should be rejected with an error when trying to set the content of a directory.", () => {
-            const uri = root.appendPath("foo");
+            const uri = root.resolve("foo");
             fs.mkdirSync(FileUri.fsPath(uri));
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.true;
             expect(fs.statSync(FileUri.fsPath(uri)).isDirectory()).to.be.true;
@@ -183,7 +188,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should be rejected with an error when trying to set the content of a file which is out-of-sync.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             fs.writeFileSync(FileUri.fsPath(uri), "foo", { encoding: "utf8" });
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.true;
             expect(fs.statSync(FileUri.fsPath(uri)).isFile()).to.be.true;
@@ -201,7 +206,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should be rejected with an error when trying to set the content when the desired encoding cannot be handled.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             fs.writeFileSync(FileUri.fsPath(uri), "foo", { encoding: "utf8" });
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.true;
             expect(fs.statSync(FileUri.fsPath(uri)).isFile()).to.be.true;
@@ -214,7 +219,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should return with a stat representing the latest state of the successfully modified file.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             fs.writeFileSync(FileUri.fsPath(uri), "foo", { encoding: "utf8" });
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.true;
             expect(fs.statSync(FileUri.fsPath(uri)).isFile()).to.be.true;
@@ -233,16 +238,16 @@ describe("NodeFileSystem", () => {
     describe("04 #move", () => {
 
         it("Should be rejected with an error if no file exists under the source location.", () => {
-            const sourceUri = root.appendPath("foo.txt");
-            const targetUri = root.appendPath("bar.txt");
+            const sourceUri = root.resolve("foo.txt");
+            const targetUri = root.resolve("bar.txt");
             expect(fs.existsSync(FileUri.fsPath(sourceUri))).to.be.false;
 
             return fileSystem.move(sourceUri.toString(), targetUri.toString()).should.eventually.be.rejectedWith(Error);
         });
 
         it("Should be rejected with an error if target exists and overwrite is not set to \'true\'.", () => {
-            const sourceUri = root.appendPath("foo.txt");
-            const targetUri = root.appendPath("bar.txt");
+            const sourceUri = root.resolve("foo.txt");
+            const targetUri = root.resolve("bar.txt");
             fs.writeFileSync(FileUri.fsPath(sourceUri), "foo");
             fs.writeFileSync(FileUri.fsPath(targetUri), "bar");
             expect(fs.statSync(FileUri.fsPath(sourceUri)).isFile()).to.be.true;
@@ -252,8 +257,8 @@ describe("NodeFileSystem", () => {
         });
 
         it("Moving a file to an empty directory. Should be rejected with an error because files cannot be moved to an existing directory locations.", () => {
-            const sourceUri = root.appendPath("foo.txt");
-            const targetUri = root.appendPath("bar");
+            const sourceUri = root.resolve("foo.txt");
+            const targetUri = root.resolve("bar");
             fs.writeFileSync(FileUri.fsPath(sourceUri), "foo");
             fs.mkdirSync(FileUri.fsPath(targetUri));
             expect(fs.statSync(FileUri.fsPath(sourceUri)).isFile()).to.be.true;
@@ -265,10 +270,10 @@ describe("NodeFileSystem", () => {
         });
 
         it("Moving a file to a non-empty directory. Should be rejected with and error because files cannot be moved to an existing directory locations.", () => {
-            const sourceUri = root.appendPath("foo.txt");
-            const targetUri = root.appendPath("bar");
-            const targetFileUri_01 = targetUri.appendPath("bar_01.txt");
-            const targetFileUri_02 = targetUri.appendPath("bar_02.txt");
+            const sourceUri = root.resolve("foo.txt");
+            const targetUri = root.resolve("bar");
+            const targetFileUri_01 = targetUri.resolve("bar_01.txt");
+            const targetFileUri_02 = targetUri.resolve("bar_02.txt");
             fs.writeFileSync(FileUri.fsPath(sourceUri), "foo");
             fs.mkdirSync(FileUri.fsPath(targetUri));
             fs.writeFileSync(FileUri.fsPath(targetFileUri_01), "bar_01");
@@ -284,8 +289,8 @@ describe("NodeFileSystem", () => {
         });
 
         it("Moving an empty directory to file. Should be rejected with an error because directories and cannot be moved to existing file locations.", () => {
-            const sourceUri = root.appendPath("foo");
-            const targetUri = root.appendPath("bar.txt");
+            const sourceUri = root.resolve("foo");
+            const targetUri = root.resolve("bar.txt");
             fs.mkdirSync(FileUri.fsPath(sourceUri));
             fs.writeFileSync(FileUri.fsPath(targetUri), "bar");
             expect(fs.statSync(FileUri.fsPath(sourceUri)).isDirectory()).to.be.true;
@@ -297,10 +302,10 @@ describe("NodeFileSystem", () => {
         });
 
         it("Moving a non-empty directory to file. Should be rejected with an error because directories cannot be moved to existing file locations.", () => {
-            const sourceUri = root.appendPath("foo");
-            const targetUri = root.appendPath("bar.txt");
-            const sourceFileUri_01 = sourceUri.appendPath("foo_01.txt");
-            const sourceFileUri_02 = sourceUri.appendPath("foo_02.txt");
+            const sourceUri = root.resolve("foo");
+            const targetUri = root.resolve("bar.txt");
+            const sourceFileUri_01 = sourceUri.resolve("foo_01.txt");
+            const sourceFileUri_02 = sourceUri.resolve("foo_02.txt");
             fs.mkdirSync(FileUri.fsPath(sourceUri));
             fs.writeFileSync(FileUri.fsPath(targetUri), "bar");
             fs.writeFileSync(FileUri.fsPath(sourceFileUri_01), "foo_01");
@@ -314,8 +319,8 @@ describe("NodeFileSystem", () => {
         });
 
         it("Moving file to file. Should overwrite the target file content and delete the source file.", () => {
-            const sourceUri = root.appendPath("foo.txt");
-            const targetUri = root.appendPath("bar.txt");
+            const sourceUri = root.resolve("foo.txt");
+            const targetUri = root.resolve("bar.txt");
             fs.writeFileSync(FileUri.fsPath(sourceUri), "foo");
             expect(fs.statSync(FileUri.fsPath(sourceUri)).isFile()).to.be.true;
             expect(fs.existsSync(FileUri.fsPath(targetUri))).to.be.false;
@@ -329,8 +334,8 @@ describe("NodeFileSystem", () => {
         });
 
         it("Moving an empty directory to an empty directory. Should remove the source directory.", () => {
-            const sourceUri = root.appendPath("foo");
-            const targetUri = root.appendPath("bar");
+            const sourceUri = root.resolve("foo");
+            const targetUri = root.resolve("bar");
             fs.mkdirSync(FileUri.fsPath(sourceUri));
             fs.mkdirSync(FileUri.fsPath(targetUri));
             expect(fs.statSync(FileUri.fsPath(sourceUri)).isDirectory()).to.be.true;
@@ -347,10 +352,10 @@ describe("NodeFileSystem", () => {
         });
 
         it("Moving an empty directory to a non-empty directory. Should be rejected because the target folder is not empty.", () => {
-            const sourceUri = root.appendPath("foo");
-            const targetUri = root.appendPath("bar");
-            const targetFileUri_01 = targetUri.appendPath("bar_01.txt");
-            const targetFileUri_02 = targetUri.appendPath("bar_02.txt");
+            const sourceUri = root.resolve("foo");
+            const targetUri = root.resolve("bar");
+            const targetFileUri_01 = targetUri.resolve("bar_01.txt");
+            const targetFileUri_02 = targetUri.resolve("bar_02.txt");
             fs.mkdirSync(FileUri.fsPath(sourceUri));
             fs.mkdirSync(FileUri.fsPath(targetUri));
             fs.writeFileSync(FileUri.fsPath(targetFileUri_01), "bar_01");
@@ -366,10 +371,10 @@ describe("NodeFileSystem", () => {
         });
 
         it("Moving a non-empty directory to an empty directory. Source folder and its content should be moved to the target location.", () => {
-            const sourceUri = root.appendPath("foo");
-            const targetUri = root.appendPath("bar");
-            const sourceFileUri_01 = sourceUri.appendPath("foo_01.txt");
-            const sourceFileUri_02 = sourceUri.appendPath("foo_02.txt");
+            const sourceUri = root.resolve("foo");
+            const targetUri = root.resolve("bar");
+            const sourceFileUri_01 = sourceUri.resolve("foo_01.txt");
+            const sourceFileUri_02 = sourceUri.resolve("foo_02.txt");
             fs.mkdirSync(FileUri.fsPath(sourceUri));
             fs.mkdirSync(FileUri.fsPath(targetUri));
             fs.writeFileSync(FileUri.fsPath(sourceFileUri_01), "foo_01");
@@ -386,18 +391,18 @@ describe("NodeFileSystem", () => {
                 expect(fs.existsSync(FileUri.fsPath(sourceUri))).to.be.false;
                 expect(fs.statSync(FileUri.fsPath(targetUri)).isDirectory()).to.be.true;
                 expect(fs.readdirSync(FileUri.fsPath(targetUri))).to.include("foo_01.txt").and.to.include("foo_02.txt");
-                expect(fs.readFileSync(FileUri.fsPath(targetUri.appendPath("foo_01.txt")), "utf8")).to.be.equal("foo_01");
-                expect(fs.readFileSync(FileUri.fsPath(targetUri.appendPath("foo_02.txt")), "utf8")).to.be.equal("foo_02");
+                expect(fs.readFileSync(FileUri.fsPath(targetUri.resolve("foo_01.txt")), "utf8")).to.be.equal("foo_01");
+                expect(fs.readFileSync(FileUri.fsPath(targetUri.resolve("foo_02.txt")), "utf8")).to.be.equal("foo_02");
             });
         });
 
         it("Moving a non-empty directory to a non-empty directory. Should be rejected because the target location is not empty.", () => {
-            const sourceUri = root.appendPath("foo");
-            const targetUri = root.appendPath("bar");
-            const sourceFileUri_01 = sourceUri.appendPath("foo_01.txt");
-            const sourceFileUri_02 = sourceUri.appendPath("foo_02.txt");
-            const targetFileUri_01 = targetUri.appendPath("bar_01.txt");
-            const targetFileUri_02 = targetUri.appendPath("bar_02.txt");
+            const sourceUri = root.resolve("foo");
+            const targetUri = root.resolve("bar");
+            const sourceFileUri_01 = sourceUri.resolve("foo_01.txt");
+            const sourceFileUri_02 = sourceUri.resolve("foo_02.txt");
+            const targetFileUri_01 = targetUri.resolve("bar_01.txt");
+            const targetFileUri_02 = targetUri.resolve("bar_02.txt");
             fs.mkdirSync(FileUri.fsPath(sourceUri));
             fs.mkdirSync(FileUri.fsPath(targetUri));
             fs.writeFileSync(FileUri.fsPath(sourceFileUri_01), "foo_01");
@@ -421,8 +426,8 @@ describe("NodeFileSystem", () => {
     describe("05 #copy", () => {
 
         it("Copy a file from non existing location. Should be rejected with an error. Nothing to copy.", () => {
-            const sourceUri = root.appendPath("foo");
-            const targetUri = root.appendPath("bar");
+            const sourceUri = root.resolve("foo");
+            const targetUri = root.resolve("bar");
             fs.mkdirSync(FileUri.fsPath(targetUri));
             expect(fs.existsSync(FileUri.fsPath(sourceUri))).to.be.false;
             expect(fs.statSync(FileUri.fsPath(targetUri)).isDirectory()).to.be.true;
@@ -431,8 +436,8 @@ describe("NodeFileSystem", () => {
         });
 
         it("Copy a file to existing location without overwrite enabled. Should be rejected with an error.", () => {
-            const sourceUri = root.appendPath("foo");
-            const targetUri = root.appendPath("bar");
+            const sourceUri = root.resolve("foo");
+            const targetUri = root.resolve("bar");
             fs.mkdirSync(FileUri.fsPath(targetUri));
             fs.mkdirSync(FileUri.fsPath(sourceUri));
             expect(fs.statSync(FileUri.fsPath(sourceUri)).isDirectory()).to.be.true;
@@ -442,8 +447,8 @@ describe("NodeFileSystem", () => {
         });
 
         it("Copy an empty directory to a non-existing location. Should return with the file stat representing the new file at the target location.", () => {
-            const sourceUri = root.appendPath("foo");
-            const targetUri = root.appendPath("bar");
+            const sourceUri = root.resolve("foo");
+            const targetUri = root.resolve("bar");
             fs.mkdirSync(FileUri.fsPath(sourceUri));
             expect(fs.statSync(FileUri.fsPath(sourceUri)).isDirectory()).to.be.true;
             expect(fs.existsSync(FileUri.fsPath(targetUri))).to.be.false;
@@ -457,8 +462,8 @@ describe("NodeFileSystem", () => {
         });
 
         it("Copy an empty directory to a non-existing, nested location. Should return with the file stat representing the new file at the target location.", () => {
-            const sourceUri = root.appendPath("foo");
-            const targetUri = root.appendPath("nested/path/to/bar");
+            const sourceUri = root.resolve("foo");
+            const targetUri = root.resolve("nested/path/to/bar");
             fs.mkdirSync(FileUri.fsPath(sourceUri));
             expect(fs.statSync(FileUri.fsPath(sourceUri)).isDirectory()).to.be.true;
             expect(fs.existsSync(FileUri.fsPath(targetUri))).to.be.false;
@@ -472,9 +477,9 @@ describe("NodeFileSystem", () => {
         });
 
         it("Copy a directory with content to a non-existing location. Should return with the file stat representing the new file at the target location.", () => {
-            const sourceUri = root.appendPath("foo");
-            const targetUri = root.appendPath("bar");
-            const subSourceUri = sourceUri.appendPath("foo_01.txt");
+            const sourceUri = root.resolve("foo");
+            const targetUri = root.resolve("bar");
+            const subSourceUri = sourceUri.resolve("foo_01.txt");
             fs.mkdirSync(FileUri.fsPath(sourceUri));
             fs.writeFileSync(FileUri.fsPath(subSourceUri), "foo");
             expect(fs.statSync(FileUri.fsPath(sourceUri)).isDirectory()).to.be.true;
@@ -490,14 +495,14 @@ describe("NodeFileSystem", () => {
                 expect(fs.readdirSync(FileUri.fsPath(sourceUri))).to.contain("foo_01.txt");
                 expect(fs.readdirSync(FileUri.fsPath(targetUri))).to.contain("foo_01.txt");
                 expect(fs.readFileSync(FileUri.fsPath(subSourceUri), "utf8")).to.be.equal("foo");
-                expect(fs.readFileSync(FileUri.fsPath(targetUri.appendPath("foo_01.txt")), "utf8")).to.be.equal("foo");
+                expect(fs.readFileSync(FileUri.fsPath(targetUri.resolve("foo_01.txt")), "utf8")).to.be.equal("foo");
             });
         });
 
         it("Copy a directory with content to a non-existing, nested location. Should return with the file stat representing the new file at the target location.", () => {
-            const sourceUri = root.appendPath("foo");
-            const targetUri = root.appendPath("nested/path/to/bar");
-            const subSourceUri = sourceUri.appendPath("foo_01.txt");
+            const sourceUri = root.resolve("foo");
+            const targetUri = root.resolve("nested/path/to/bar");
+            const subSourceUri = sourceUri.resolve("foo_01.txt");
             fs.mkdirSync(FileUri.fsPath(sourceUri));
             fs.writeFileSync(FileUri.fsPath(subSourceUri), "foo");
             expect(fs.statSync(FileUri.fsPath(sourceUri)).isDirectory()).to.be.true;
@@ -513,47 +518,7 @@ describe("NodeFileSystem", () => {
                 expect(fs.readdirSync(FileUri.fsPath(sourceUri))).to.contain("foo_01.txt");
                 expect(fs.readdirSync(FileUri.fsPath(targetUri))).to.contain("foo_01.txt");
                 expect(fs.readFileSync(FileUri.fsPath(subSourceUri), "utf8")).to.be.equal("foo");
-                expect(fs.readFileSync(FileUri.fsPath(targetUri.appendPath("foo_01.txt")), "utf8")).to.be.equal("foo");
-            });
-        });
-
-    });
-
-    describe("06 #getWorkspaceRoot", () => {
-
-        it("Should be return with the stat of the root. The root stat has information of its direct descendants but not the children of the descendants.", () => {
-            const uri_1 = root.appendPath("foo");
-            const uri_2 = root.appendPath("bar");
-            const uri_1_01 = uri_1.appendPath("foo_01.txt");
-            const uri_1_02 = uri_1.appendPath("foo_02.txt");
-            const uri_2_01 = uri_2.appendPath("bar_01.txt");
-            const uri_2_02 = uri_2.appendPath("bar_02.txt");
-            fs.mkdirSync(FileUri.fsPath(uri_1));
-            fs.mkdirSync(FileUri.fsPath(uri_2));
-            fs.writeFileSync(FileUri.fsPath(uri_1_01), "foo_01");
-            fs.writeFileSync(FileUri.fsPath(uri_1_02), "foo_02");
-            fs.writeFileSync(FileUri.fsPath(uri_2_01), "bar_01");
-            fs.writeFileSync(FileUri.fsPath(uri_2_02), "bar_02");
-            expect(fs.statSync(FileUri.fsPath(uri_1)).isDirectory()).to.be.true;
-            expect(fs.statSync(FileUri.fsPath(uri_2)).isDirectory()).to.be.true;
-            expect(fs.readFileSync(FileUri.fsPath(uri_1_01), "utf8")).to.be.equal("foo_01");
-            expect(fs.readFileSync(FileUri.fsPath(uri_1_02), "utf8")).to.be.equal("foo_02");
-            expect(fs.readFileSync(FileUri.fsPath(uri_2_01), "utf8")).to.be.equal("bar_01");
-            expect(fs.readFileSync(FileUri.fsPath(uri_2_02), "utf8")).to.be.equal("bar_02");
-            expect(fs.readdirSync(FileUri.fsPath(uri_1))).to.include("foo_01.txt").and.to.include("foo_02.txt");
-            expect(fs.readdirSync(FileUri.fsPath(uri_2))).to.include("bar_01.txt").and.to.include("bar_02.txt");
-
-            return fileSystem.getWorkspaceRoot().then(stat => {
-                expect(stat).to.be.an("object");
-                expect(stat).to.have.property("uri").that.equals(root.toString());
-                expect(stat).to.have.property("hasChildren").that.be.true;
-                expect(stat).to.have.property("children").that.is.not.undefined;
-                expect(stat).to.have.property("children").that.has.lengthOf(2);
-                expect(stat.children!.map(childStat => childStat.uri)).to.contain(uri_1.toString()).and.contain(uri_2.toString());
-                expect(stat.children!.find(childStat => childStat.uri === uri_1.toString())).to.be.not.undefined;
-                expect(stat.children!.find(childStat => childStat.uri === uri_2.toString())).to.be.not.undefined;
-                expect(stat.children!.find(childStat => childStat.uri === uri_1.toString())!.children).to.be.undefined;
-                expect(stat.children!.find(childStat => childStat.uri === uri_2.toString())!.children).to.be.undefined;
+                expect(fs.readFileSync(FileUri.fsPath(targetUri.resolve("foo_01.txt")), "utf8")).to.be.equal("foo");
             });
         });
 
@@ -562,7 +527,7 @@ describe("NodeFileSystem", () => {
     describe("07 #createFile", () => {
 
         it("Should be rejected with an error if a file already exists with the given URI.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             fs.writeFileSync(FileUri.fsPath(uri), "foo");
             expect(fs.statSync(FileUri.fsPath(uri)).isFile()).to.be.true;
 
@@ -570,14 +535,14 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should be rejected with an error if the encoding is given but cannot be handled.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             return fileSystem.createFile(uri.toString(), { encoding: "unknownEncoding" }).should.be.eventually.rejectedWith(Error);
         });
 
         it("Should create an empty file without any contents by default.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             return fileSystem.createFile(uri.toString()).then(stat => {
@@ -590,7 +555,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should create a file with the desired content.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             return fileSystem.createFile(uri.toString(), { content: "foo" }).then(stat => {
@@ -603,7 +568,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should create a file with the desired content into a non-existing, nested location.", () => {
-            const uri = root.appendPath("foo/bar/baz.txt");
+            const uri = root.resolve("foo/bar/baz.txt");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             return fileSystem.createFile(uri.toString(), { content: "foo" }).then(stat => {
@@ -616,7 +581,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should create a file with the desired content and encoding.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             return fileSystem.createFile(uri.toString(), { content: "foo", encoding: "utf8" }).then(stat => {
@@ -633,7 +598,7 @@ describe("NodeFileSystem", () => {
     describe("08 #createFolder", () => {
 
         it("Should be rejected with an error if a directory already exist under the desired URI.", () => {
-            const uri = root.appendPath("foo");
+            const uri = root.resolve("foo");
             fs.mkdirSync(FileUri.fsPath(uri));
             expect(fs.statSync(FileUri.fsPath(uri)).isDirectory()).to.be.true;
 
@@ -641,7 +606,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should create a directory and return with the stat object on successful directory creation.", () => {
-            const uri = root.appendPath("foo");
+            const uri = root.resolve("foo");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             return fileSystem.createFolder(uri.toString()).then(stat => {
@@ -653,7 +618,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should create a directory and return with the stat object on successful directory creation.", () => {
-            const uri = root.appendPath("foo/bar");
+            const uri = root.resolve("foo/bar");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             return fileSystem.createFolder(uri.toString()).then(stat => {
@@ -669,7 +634,7 @@ describe("NodeFileSystem", () => {
     describe("09 #touch", () => {
 
         it("Should create a new file if it does not exist yet.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             return fileSystem.touchFile(uri.toString()).then(stat => {
@@ -680,7 +645,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should update the modification timestamp on an existing file.", done => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             fs.writeFileSync(FileUri.fsPath(uri), "foo");
             expect(fs.statSync(FileUri.fsPath(uri)).isFile()).to.be.true;
 
@@ -709,14 +674,14 @@ describe("NodeFileSystem", () => {
     describe("#10 delete", () => {
 
         it("Should be rejected when the file to delete does not exist.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             return fileSystem.delete(uri.toString(), { moveToTrash: false }).should.be.eventually.rejectedWith(Error);
         });
 
         it("Should delete the file.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             fs.writeFileSync(FileUri.fsPath(uri), "foo");
             expect(fs.readFileSync(FileUri.fsPath(uri), "utf8")).to.be.equal("foo");
 
@@ -726,7 +691,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should delete a directory without content.", () => {
-            const uri = root.appendPath("foo");
+            const uri = root.resolve("foo");
             fs.mkdirSync(FileUri.fsPath(uri));
             expect(fs.statSync(FileUri.fsPath(uri)).isDirectory()).to.be.true;
 
@@ -736,8 +701,8 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should delete a directory with all its content.", () => {
-            const uri = root.appendPath("foo");
-            const subUri = uri.appendPath("bar.txt");
+            const uri = root.resolve("foo");
+            const subUri = uri.resolve("bar.txt");
             fs.mkdirSync(FileUri.fsPath(uri));
             fs.writeFileSync(FileUri.fsPath(subUri), "bar");
             expect(fs.statSync(FileUri.fsPath(uri)).isDirectory()).to.be.true;
@@ -754,14 +719,14 @@ describe("NodeFileSystem", () => {
     describe("#11 getEncoding", () => {
 
         it("Should be rejected with an error if no file exists under the given URI.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             expect(fs.existsSync(FileUri.fsPath(uri))).to.be.false;
 
             return fileSystem.getEncoding(uri.toString()).should.be.eventually.rejectedWith(Error);
         });
 
         it("Should be rejected with an error if the URI points to a directory instead of a file.", () => {
-            const uri = root.appendPath("foo");
+            const uri = root.resolve("foo");
             fs.mkdirSync(FileUri.fsPath(uri));
             expect(fs.statSync(FileUri.fsPath(uri)).isDirectory()).to.be.true;
 
@@ -769,7 +734,7 @@ describe("NodeFileSystem", () => {
         });
 
         it("Should return with the encoding of the file.", () => {
-            const uri = root.appendPath("foo.txt");
+            const uri = root.resolve("foo.txt");
             fs.writeFileSync(FileUri.fsPath(uri), "foo");
             expect(fs.statSync(FileUri.fsPath(uri)).isFile()).to.be.true;
 
@@ -778,56 +743,55 @@ describe("NodeFileSystem", () => {
 
     });
 
-    describe("#12 constructor", () => {
-
-        it("Should throw an exception if the workspace root does not exist.", () => {
-            return expect(() => new FileSystemNode(FileUri.create("some/missing/path"))).to.throw(Error);
-        });
-
-    });
-
     describe("#13 watchFileChanges", () => {
 
         it("Should receive file changes events from in the workspace by default.", function (done) {
             this.timeout(4000);
-            let expectedEvents = [
-                new FileChange(root.toString(), FileChangeType.ADDED),
-                new FileChange(root.appendPath("foo").toString(), FileChangeType.ADDED),
-                new FileChange(root.appendPath("foo").appendPath("bar").toString(), FileChangeType.ADDED),
-                new FileChange(root.appendPath("foo").appendPath("bar").appendPath("baz.txt").toString(), FileChangeType.ADDED)
+            const type = FileChangeType.ADDED;
+            const expectedChanges: FileChange[] = [
+                { uri: root.resolve("foo"), type },
+                { uri: root.withPath(root.path.join('foo', 'bar')), type },
+                { uri: root.withPath(root.path.join('foo', 'bar', 'baz.txt')), type }
             ];
-            const fileSystem = createFileSystem();
-            const client: FileSystemClient = {
-                onFileChanges(event: FileChangesEvent) {
-                    const index = expectedEvents.findIndex((value, index, obj) => {
-                        return event.changes.length === 1 && event.changes[0].equals(value);
-                    });
-                    if (index >= 0) {
-                        expectedEvents.splice(index, 1);
-                    }
-                    if (expectedEvents.length === 0) {
-                        (<FileSystemNode>fileSystem).setClient(undefined);
-                        fileSystem.dispose();
-                        done();
-                    }
-                }
-            };
-            (<FileSystemNode>fileSystem).setClient(client);
-            fs.mkdirSync(FileUri.fsPath(root.appendPath("foo")));
-            expect(fs.statSync(FileUri.fsPath(root.appendPath("foo"))).isDirectory()).to.be.true;
-            fs.mkdirSync(FileUri.fsPath(root.appendPath("foo").appendPath("bar")));
-            expect(fs.statSync(FileUri.fsPath(root.appendPath("foo").appendPath("bar"))).isDirectory()).to.be.true;
-            fs.writeFileSync(FileUri.fsPath(root.appendPath("foo").appendPath("bar").appendPath("baz.txt")), "baz");
-            expect(fs.readFileSync(FileUri.fsPath(root.appendPath("foo").appendPath("bar").appendPath("baz.txt")), "utf8")).to.be.equal("baz");
-            sleep(3000).then(() => {
-                expect(expectedEvents).to.be.empty;
+            const actualChanges: FileChange[] = [];
+            watcher.onFilesChanged(changes =>
+                actualChanges.push(...changes)
+            );
+            watcher.watchFileChanges(root).then(() => {
+                fs.mkdirSync(FileUri.fsPath(root.resolve("foo")));
+                expect(fs.statSync(FileUri.fsPath(root.resolve("foo"))).isDirectory()).to.be.true;
+                fs.mkdirSync(FileUri.fsPath(root.resolve("foo").resolve("bar")));
+                expect(fs.statSync(FileUri.fsPath(root.resolve("foo").resolve("bar"))).isDirectory()).to.be.true;
+                fs.writeFileSync(FileUri.fsPath(root.resolve("foo").resolve("bar").resolve("baz.txt")), "baz");
+                expect(fs.readFileSync(FileUri.fsPath(root.resolve("foo").resolve("bar").resolve("baz.txt")), "utf8")).to.be.equal("baz");
             });
+            setTimeout(() => {
+                assert.deepEqual(expectedChanges, actualChanges);
+                done();
+            }, 2000);
         });
 
     });
 
-    function createFileSystem(uri: URI = root): FileSystem {
-        return new FileSystemNode(uri);
+    function createFileSystem(): FileSystem {
+        return new FileSystemNode();
+    }
+
+    function createFileSystemWatcher(): FileSystemWatcher {
+        const logger = new Proxy<Logger>({} as any, {
+            get: (target, name) => () => {
+                if (name.toString().startsWith('is')) {
+                    return Promise.resolve(false);
+                }
+                if (name.toString().startsWith('if')) {
+                    return new Promise(resolve => { });
+                }
+            }
+        });
+        const listener = new FileSystemWatcherClientListener();
+        const server = new ChokidarFileSystemWatcherServer(logger);
+        server.setClient(listener);
+        return new FileSystemWatcher(server, listener);
     }
 
     function sleep(time: number) {
