@@ -5,66 +5,40 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-import * as fs from "fs-extra";
-import * as touch from "touch";
-import * as path from "path";
-import { FileUri } from "../../application/node/file-uri";
+import * as paths from 'path';
+import * as fs from 'fs-extra';
+import * as touch from 'touch';
+import { injectable, inject, optional } from "inversify";
 import URI from "../../application/common/uri";
-import { WatchOptions, FSWatcher } from "chokidar";
-import { FileStat, FileSystem, FileSystemClient, FileChange, FileChangeType, FileChangesEvent } from "../common/filesystem";
+import { FileUri } from "../../application/node";
+import { FileStat, FileSystem } from "../common/filesystem";
 
 type MvOptions = { mkdirp?: boolean, clobber?: boolean, limit?: number };
 
 const trash: (paths: Iterable<string>) => Promise<void> = require("trash");
-const chokidar: { watch(paths: string | string[], options?: WatchOptions): FSWatcher } = require("chokidar");
 const mv: (sourcePath: string, targetPath: string, options?: MvOptions, cb?: (error: NodeJS.ErrnoException) => void) => void = require("mv");
 
-type EventType =
-    "all" |
-    "add" |
-    "addDir" |
-    "unlink" |
-    "unlinkDir" |
-    "change" |
-    "error";
+@injectable()
+export class FileSystemNodeOptions {
+    encoding: string;
+    recursive: boolean;
+    overwrite: boolean;
+    moveToTrash: true;
 
-export class FileSystemNode implements FileSystem {
-
-    protected client: FileSystemClient | undefined;
-    private watcher: FSWatcher;
-
-    constructor(protected rootUri: URI, protected defaults: FileSystem.Configuration = {
+    public static default: FileSystemNodeOptions = {
         encoding: "utf8",
         overwrite: false,
         recursive: true,
-        moveToTrash: true,
-    }) {
-
-        const stat = this.doGetStat(this.rootUri, 0);
-        if (!stat) {
-            throw new Error(`File system root cannot be located under ${this.rootUri}.`);
-        }
-        if (!stat.isDirectory) {
-            throw new Error(`File system root should point to a directory location. URI: ${this.rootUri}.`);
-        }
-        const rootPath = FileUri.fsPath(this.rootUri);
-        this.watcher = chokidar.watch(rootPath).on("all", (eventType: EventType, filename: string) => {
-            if (this.client) {
-                const changeType = this.getFileChangeType(eventType);
-                if (changeType !== undefined && typeof filename === "string") {
-                    const segments = path.normalize(path.relative(rootPath, filename));
-                    const changedUri = segments === "." ? rootUri : FileUri.create(path.resolve(rootPath, segments));
-                    const change = new FileChange(changedUri.toString(), changeType);
-                    const event = new FileChangesEvent([change]);
-                    this.client.onFileChanges(event);
-                }
-            }
-        });
+        moveToTrash: true
     }
+}
 
-    setClient(client: FileSystemClient | undefined) {
-        this.client = client;
-    }
+@injectable()
+export class FileSystemNode implements FileSystem {
+
+    constructor(
+        @inject(FileSystemNodeOptions) @optional() protected readonly options: FileSystemNodeOptions = FileSystemNodeOptions.default
+    ) { }
 
     getFileStat(uriAsString: string): Promise<FileStat> {
         const uri = new URI(uriAsString);
@@ -120,7 +94,7 @@ export class FileSystemNode implements FileSystem {
                 return reject(new Error(`File is out of sync. URI: ${file.uri}. Expected size: ${stat.size}. Actual size: ${file.size}.`));
             }
             const encoding = this.doGetEncoding(options);
-            fs.writeFile(FileUri.fsPath(_uri), content, encoding, error => {
+            fs.writeFile(FileUri.fsPath(_uri), content, { encoding }, error => {
                 if (error) {
                     return reject(error);
                 }
@@ -303,28 +277,6 @@ export class FileSystemNode implements FileSystem {
         });
     }
 
-    watchFileChanges(uri: string): Promise<void> {
-        return new Promise<void>((resolve, reject) => {
-            const _uri = new URI(uri);
-            const stat = this.doGetStat(_uri, 0);
-            if (!stat) {
-                return reject(new Error(`File does not exist under ${uri}.`));
-            }
-            this.watcher.add(FileUri.fsPath(_uri));
-        });
-    }
-
-    unwatchFileChanges(uri: string): Promise<void> {
-        return new Promise<void>((resolve, reject) => {
-            const _uri = new URI(uri);
-            const stat = this.doGetStat(_uri, 0);
-            if (!stat) {
-                return reject(new Error(`File does not exist under ${uri}.`));
-            }
-            this.watcher.unwatch(FileUri.fsPath(_uri));
-        });
-    }
-
     getEncoding(uri: string): Promise<string> {
         return new Promise<string>((resolve, reject) => {
             const _uri = new URI(uri);
@@ -335,106 +287,102 @@ export class FileSystemNode implements FileSystem {
             if (stat.isDirectory) {
                 return reject(new Error(`Cannot get the encoding of a director. URI: ${uri}.`));
             }
-            return resolve(this.defaults.encoding);
+            return resolve(this.options.encoding);
         });
     }
 
-    getWorkspaceRoot(): Promise<FileStat> {
-        return new Promise<FileStat>((resolve, reject) => {
-            const stat = this.doGetStat(this.rootUri, 1);
-            if (!stat) {
-                return reject(new Error(`Cannot locate workspace root under ${this.rootUri}.`));
-            }
-            resolve(stat);
-        });
+    getRoots(): Promise<FileStat[]> {
+        const cwdRoot = paths.parse(process.cwd()).root;
+        const rootUri = FileUri.create(cwdRoot);
+        const root = this.doGetStat(rootUri, 1);
+        if (root) {
+            return Promise.resolve([root]);
+        }
+        console.error(`Cannot locate the file system root under ${rootUri}.`);
+        return Promise.resolve([]);
     }
 
     dispose(): void {
-        this.watcher.close();
     }
 
     protected doGetStat(uri: URI, depth: number): FileStat | undefined {
         const path = FileUri.fsPath(uri);
         try {
-            const stat = fs.statSync(path);
-            if (stat.isDirectory()) {
-                const files = fs.readdirSync(path);
-                let children: FileStat[] | undefined = undefined;
-                if (depth > 0) {
-                    children = [];
-                    files.map(file => uri.appendPath(file)).forEach(childURI => {
-                        const child = this.doGetStat(childURI, depth - 1);
-                        if (child) {
-                            children!.push(child);
-                        }
-                    });
-                }
-                return {
-                    uri: uri.toString(),
-                    lastModification: stat.mtime.getTime(),
-                    isDirectory: true,
-                    hasChildren: files.length > 0,
-                    children
-                };
-            } else {
-                return {
-                    uri: uri.toString(),
-                    lastModification: stat.mtime.getTime(),
-                    isDirectory: false,
-                    size: stat.size
-                };
+            const stats = fs.statSync(path);
+            if (stats.isDirectory()) {
+                return this.doCreateDirectoryStat(uri, path, stats, depth);
             }
+            return this.doCreateFileStat(uri, stats);
         } catch (error) {
-            if (isErrnoException(error) && error.code === "ENOENT") {
-                return undefined;
+            if (isErrnoException(error)) {
+                if (error.code === "ENOENT" || error.code === "EACCES") {
+                    return undefined;
+                }
             }
             throw error;
         }
     }
 
+    protected doCreateFileStat(uri: URI, stat: fs.Stats): FileStat {
+        return {
+            uri: uri.toString(),
+            lastModification: stat.mtime.getTime(),
+            isDirectory: false,
+            size: stat.size
+        };
+    }
+
+    protected doCreateDirectoryStat(uri: URI, path: string, stat: fs.Stats, depth: number): FileStat {
+        const files = fs.readdirSync(path);
+        const hasChildren = files.length > 0;
+        const children = hasChildren && depth > 0 ? this.doGetChildren(uri, files, depth) : undefined;
+        return {
+            uri: uri.toString(),
+            lastModification: stat.mtime.getTime(),
+            isDirectory: true,
+            hasChildren,
+            children
+        };
+    }
+
+    protected doGetChildren(uri: URI, files: string[], depth: number): FileStat[] {
+        const children = [];
+        for (const file of files) {
+            const childUri = uri.resolve(file);
+            const child = this.doGetStat(childUri, depth - 1);
+            if (child) {
+                children.push(child);
+            }
+        }
+        return children;
+    }
+
     protected doGetEncoding(option?: { encoding?: string }): string {
         return option && typeof (option.encoding) !== "undefined"
             ? option.encoding
-            : this.defaults.encoding;
+            : this.options.encoding;
     }
 
     protected doGetOverwrite(option?: { overwrite?: boolean }): boolean {
         return option && typeof (option.overwrite) !== "undefined"
             ? option.overwrite
-            : this.defaults.overwrite;
+            : this.options.overwrite;
     }
 
     protected doGetRecursive(option?: { recursive?: boolean }): boolean {
         return option && typeof (option.recursive) !== "undefined"
             ? option.recursive
-            : this.defaults.recursive;
+            : this.options.recursive;
     }
 
     protected doGetMoveToTrash(option?: { moveToTrash?: boolean }): boolean {
         return option && typeof (option.moveToTrash) !== "undefined"
             ? option.moveToTrash
-            : this.defaults.moveToTrash;
+            : this.options.moveToTrash;
     }
 
     protected doGetContent(option?: { content?: string }): string {
         return (option && option.content) || "";
-    }
-
-    private getFileChangeType(eventType: EventType): FileChangeType | undefined {
-        switch (eventType) {
-            case "add":
-            case "addDir":
-                return FileChangeType.ADDED;
-            case "unlink":
-            case "unlinkDir":
-                return FileChangeType.DELETED;
-            case "change":
-                return FileChangeType.UPDATED;
-            case "error":
-                return undefined;
-            default:
-                throw new Error(`Unexpected file change event type: ${event}.`);
-        }
     }
 
 }
