@@ -9,4 +9,4 @@ export * from '../common';
 export * from './language-client-factory';
 export * from './language-client-contribution';
 export * from './languages-frontend-contribution';
-export * from './frontend-languages-module';
+export * from './languages-frontend-module';
