@@ -5,6 +5,6 @@
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
  */
 
-export { FileNavigatorModel } from './navigator-model';
-export { FileNavigatorWidget } from './navigator-widget';
-export * from './navigator-module';
+export * from './navigator-model';
+export * from './navigator-widget';
+export * from './navigator-frontend-module';
