@@ -6,79 +6,25 @@
  */
 
 import { injectable, inject } from "inversify";
-import { FileSystem, FileStat, FileChangesEvent, FileChangeType } from "../../filesystem/common/filesystem";
-import { FileSystemWatcher } from "../../filesystem/common/filesystem-watcher";
-import { UriSelection } from "../../filesystem/common/filesystem-selection";
-import { SelectionService } from '../../application/common';
-import { OpenerService, open } from "../../application/browser";
-import {
-    ITree,
-    ITreeSelectionService,
-    ITreeExpansionService,
-    ITreeNode,
-    ICompositeTreeNode,
-    IExpandableTreeNode,
-    TreeModel,
-    Tree
-} from "./tree";
-import { ISelectableTreeNode } from "./tree/tree-selection";
-import URI from '../../application/common/uri';
+import { OpenerService, open, ITreeNode } from "../../application/browser";
+import { FileNode, FileTreeModel, FileTreeServices } from "../../filesystem/browser";
+import { FileNavigatorTree } from "./navigator-tree";
 
 @injectable()
-export class FileNavigatorModel extends TreeModel {
+export class FileNavigatorServices extends FileTreeServices {
+    @inject(OpenerService) readonly openerService: OpenerService;
+}
+
+@injectable()
+export class FileNavigatorModel extends FileTreeModel {
+
+    protected readonly openerService: OpenerService;
 
     constructor(
-        @inject(FileSystem) protected readonly fileSystem: FileSystem,
-        @inject(FileSystemWatcher) watcher: FileSystemWatcher,
-        @inject(OpenerService) protected readonly openerService: OpenerService,
-        @inject(ITree) tree: ITree,
-        @inject(ITreeSelectionService) selection: ITreeSelectionService,
-        @inject(ITreeExpansionService) expansion: ITreeExpansionService,
-        @inject(SelectionService) selectionService: SelectionService
+        @inject(FileNavigatorTree) protected readonly tree: FileNavigatorTree,
+        @inject(FileNavigatorServices) services: FileNavigatorServices
     ) {
-        super(tree, selection, expansion);
-        this.toDispose.push(watcher.onFileChanges(event => this.onFileChanges(event)));
-        this.toDispose.push(selection.onSelectionChanged(selection => selectionService.selection = selection));
-    }
-
-    get selectedFileStatNode(): Readonly<FileStatNode> | undefined {
-        const selectedNode = this.selectedNode;
-        if (FileStatNode.is(selectedNode)) {
-            return selectedNode;
-        }
-        return undefined;
-    }
-
-    protected onFileChanges(event: FileChangesEvent): void {
-        const affectedNodes = this.getAffectedNodes(event);
-        if (affectedNodes.length !== 0) {
-            affectedNodes.forEach(node => this.refresh(node));
-        } else if (this.isRootAffected(event)) {
-            this.refresh();
-        }
-    }
-
-    protected isRootAffected(event: FileChangesEvent): boolean {
-        const root = this.root;
-        if (FileStatNode.is(root)) {
-            return event.changes.some(change =>
-                change.type < FileChangeType.DELETED && change.uri === root.fileStat.uri
-            );
-        }
-        return false;
-    }
-
-    protected getAffectedNodes(event: FileChangesEvent): ICompositeTreeNode[] {
-        const nodes: DirNode[] = [];
-        for (const change of event.changes) {
-            const uri = change.uri;
-            const id = change.type > FileChangeType.UPDATED ? new URI(uri).parent.toString() : uri;
-            const node = this.getNode(id);
-            if (DirNode.is(node) && node.expanded) {
-                nodes.push(node);
-            }
-        }
-        return nodes;
+        super(tree, services);
     }
 
     protected doOpenNode(node: ITreeNode): void {
@@ -88,116 +34,5 @@ export class FileNavigatorModel extends TreeModel {
             super.doOpenNode(node);
         }
     }
-}
 
-@injectable()
-export class FileNavigatorTree extends Tree {
-
-    constructor( @inject(FileSystem) protected readonly fileSystem: FileSystem) {
-        super();
-        this.fileSystem.getWorkspaceRoot().then(fileStat => {
-            this.root = this.createRootNode(fileStat);
-        });
-    }
-
-    protected createRootNode(fileStat: FileStat): DirNode {
-        const uri = new URI(fileStat.uri)
-        const id = fileStat.uri;
-        return {
-            id, uri, fileStat,
-            name: '/',
-            visible: false,
-            parent: undefined,
-            children: [],
-            expanded: true,
-            selected: false
-        }
-    }
-
-    resolveChildren(parent: ICompositeTreeNode): Promise<ITreeNode[]> {
-        if (FileStatNode.is(parent)) {
-            return this.resolveFileStat(parent).then(fileStat =>
-                this.toNodes(fileStat, parent)
-            )
-        }
-        return super.resolveChildren(parent);
-    }
-
-    protected resolveFileStat(node: FileStatNode): Promise<FileStat> {
-        return this.fileSystem.getFileStat(node.fileStat.uri).then(fileStat => {
-            node.fileStat = fileStat;
-            return fileStat;
-        });
-    }
-
-    protected toNodes(fileStat: FileStat, parent: ICompositeTreeNode): ITreeNode[] {
-        if (!fileStat.children) {
-            return [];
-        }
-        return fileStat.children.map(child =>
-            this.toNode(child, parent)
-        ).sort(DirNode.compare);
-    }
-
-    protected toNode(fileStat: FileStat, parent: ICompositeTreeNode): FileNode | DirNode {
-        const uri = new URI(fileStat.uri)
-        const id = fileStat.uri;
-        const node = this.getNode(id);
-        if (fileStat.isDirectory) {
-            if (DirNode.is(node)) {
-                node.fileStat = fileStat;
-                return node;
-            }
-            const name = uri.lastSegment;
-            return <DirNode>{
-                id, uri, fileStat, name, parent,
-                expanded: false,
-                selected: false,
-                children: []
-            }
-        }
-        if (FileNode.is(node)) {
-            node.fileStat = fileStat;
-            return node;
-        }
-        const name = uri.lastSegment;
-        return <FileNode>{
-            id, uri, fileStat, name, parent,
-            selected: false
-        }
-    }
-
-}
-
-export interface FileStatNode extends ISelectableTreeNode, UriSelection {
-    fileStat: FileStat;
-}
-export namespace FileStatNode {
-    export function is(node: ITreeNode | undefined): node is FileStatNode {
-        return !!node && 'fileStat' in node;
-    }
-}
-
-export type FileNode = FileStatNode;
-export namespace FileNode {
-    export function is(node: ITreeNode | undefined): node is FileNode {
-        return FileStatNode.is(node) && !IExpandableTreeNode.is(node);
-    }
-}
-
-export type DirNode = FileStatNode & IExpandableTreeNode;
-export namespace DirNode {
-    export function is(node: ITreeNode | undefined): node is DirNode {
-        return FileStatNode.is(node) && IExpandableTreeNode.is(node);
-    }
-
-    export function compare(node: ITreeNode, node2: ITreeNode): number {
-        return DirNode.dirCompare(node, node2) || node.name.localeCompare(node2.name);
-    }
-
-    export function dirCompare(node: ITreeNode, node2: ITreeNode): number {
-        const a = DirNode.is(node) ? 1 : 0;
-        const b = DirNode.is(node2) ? 1 : 0;
-        return b - a;
-    }
 }
